+++ CHANGES.md
@@ -7,6 +7,8 @@ Unreleased
 * atdgen: Add support for `<json open_enum>` in Melange (#401)
 * atdcpp: Initial Release (#404)
 * atdcpp: Use `double` c++ type as default floating point type (#411)
+* atdgen: Fix JSON I/O for inline records (#419)
+
 
 2.15.0 (2023-10-26)
 -------------------